[refactor] refactoring the code base for snark-verifier
