+++ zkevm/Cargo.toml
@@ -6,9 +6,9 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-halo2-snark-aggregator-circuit = { git = "https://github.com/scroll-tech/halo2-snark-aggregator", branch = "scroll-dev-0220" }
-halo2-snark-aggregator-solidity = { git = "https://github.com/scroll-tech/halo2-snark-aggregator", branch = "scroll-dev-0220" }
-halo2-snark-aggregator-api = { git = "https://github.com/scroll-tech/halo2-snark-aggregator", branch = "scroll-dev-0220" }
+# halo2-snark-aggregator-circuit = { git = "https://github.com/scroll-tech/halo2-snark-aggregator", branch = "scroll-dev-0220" }
+# halo2-snark-aggregator-solidity = { git = "https://github.com/scroll-tech/halo2-snark-aggregator", branch = "scroll-dev-0220" }
+# halo2-snark-aggregator-api = { git = "https://github.com/scroll-tech/halo2-snark-aggregator", branch = "scroll-dev-0220" }
 
 halo2_proofs = { git = "https://github.com/privacy-scaling-explorations/halo2.git", tag = "v2022_09_10" }
 #halo2_proofs = { git = "https://github.com/privacy-scaling-explorations/halo2.git", tag = "v2022_09_10", features = ["phase-check"] }
@@ -49,8 +49,8 @@ itertools = "0.10.5"
 procfs = "0.13.0"
 
 [features]
-default = []
-#default = ["prove_verify"]
+# default = []
+default = ["prove_verify"]
 prove_verify = []
 
 [dev-dependencies]
