+++ zkevm/src/verifier.rs
@@ -2,7 +2,7 @@ use std::collections::HashMap;
 use std::io::Cursor;
 
 use crate::circuit::{TargetCircuit, AGG_DEGREE, DEGREE};
-use crate::io::{deserialize_fr_matrix, load_instances};
+use crate::io::load_instances;
 use crate::prover::{AggCircuitProof, TargetCircuitProof};
 use crate::utils::load_params;
 use halo2_proofs::halo2curves::bn256::{Bn256, Fr, G1Affine};
@@ -10,11 +10,14 @@ use halo2_proofs::plonk::VerifyingKey;
 use halo2_proofs::plonk::{keygen_vk, verify_proof};
 use halo2_proofs::poly::commitment::ParamsProver;
 use halo2_proofs::poly::kzg::commitment::ParamsKZG;
-use halo2_proofs::poly::kzg::multiopen::VerifierGWC;
-use halo2_proofs::poly::kzg::strategy::SingleStrategy;
-use halo2_proofs::transcript::{Challenge255, PoseidonRead};
-use halo2_snark_aggregator_api::transcript::sha::ShaRead;
-use halo2_snark_aggregator_circuit::verify_circuit::Halo2VerifierCircuit;
+use halo2_proofs::poly::kzg::multiopen::VerifierSHPLONK;
+use halo2_proofs::poly::kzg::strategy::{AccumulatorStrategy, SingleStrategy};
+use halo2_proofs::poly::VerificationStrategy;
+use halo2_proofs::transcript::{TranscriptReadBuffer};
+use snark_verifier::system::halo2::transcript::evm::EvmTranscript;
+use snark_verifier_sdk::evm::evm_verify;
+use snark_verifier_sdk::halo2::PoseidonTranscript;
+use snark_verifier_sdk::halo2::aggregation::AggregationCircuit;
 
 pub struct Verifier {
     params: ParamsKZG<Bn256>,
@@ -35,7 +38,7 @@ impl Verifier {
             log::error!("Verifier should better have raw_agg_vk to check consistency");
         }
         let agg_vk = raw_agg_vk.as_ref().map(|k| {
-            VerifyingKey::<G1Affine>::read::<_, Halo2VerifierCircuit<'_, Bn256>>(
+            VerifyingKey::<G1Affine>::read::<_, AggregationCircuit>(
                 &mut Cursor::new(&k),
                 halo2_proofs::SerdeFormat::Processed,
             )
@@ -65,19 +68,19 @@ impl Verifier {
         Self::from_params(params, agg_params, agg_vk)
     }
 
-    pub fn verify_agg_circuit_proof(&self, proof: AggCircuitProof) -> anyhow::Result<()> {
-        if let Some(raw_agg_vk) = &self.raw_agg_vk {
-            if &proof.vk != raw_agg_vk {
-                log::error!("vk provided in proof != vk in verifier");
-            }
-        }
+    pub fn verify_agg_circuit_proof(&self, proof: AggCircuitProof) -> anyhow::Result<bool> {
+        let mut transcript = TranscriptReadBuffer::<_, G1Affine, _>::init(proof.proof.as_slice());
+
+        let vk = match self.agg_vk.clone() {
+            Some(p) => p,
+            None => panic!("aggregation verification key is not found"),
+        };
+
+        // deserialize instances
         let verify_circuit_instance: Vec<Vec<Vec<Fr>>> = {
             let instance = proof.instance;
             load_instances(&instance)
         };
-        let params = self.agg_params.verifier_params();
-        let strategy = SingleStrategy::new(params);
-
         let verify_circuit_instance1: Vec<Vec<&[Fr]>> = verify_circuit_instance
             .iter()
             .map(|x| x.iter().map(|y| &y[..]).collect())
@@ -85,45 +88,37 @@ impl Verifier {
         let verify_circuit_instance2: Vec<&[&[Fr]]> =
             verify_circuit_instance1.iter().map(|x| &x[..]).collect();
 
-        let mut transcript = ShaRead::<_, _, Challenge255<_>, sha2::Sha256>::init(&proof.proof[..]);
-
-        // TODO better way to do this?
-        let vk_in_proof = VerifyingKey::<G1Affine>::read::<_, Halo2VerifierCircuit<'_, Bn256>>(
-            &mut Cursor::new(&proof.vk),
-            halo2_proofs::SerdeFormat::Processed,
-        )
-        .unwrap();
-        verify_proof::<_, VerifierGWC<_>, _, _, _>(
-            params,
-            self.agg_vk.as_ref().unwrap_or(&vk_in_proof),
-            strategy,
-            &verify_circuit_instance2[..],
-            &mut transcript,
-        )?;
-        Ok(())
+        Ok(VerificationStrategy::<_, VerifierSHPLONK<Bn256>>::finalize(
+            verify_proof::<_, VerifierSHPLONK<Bn256>, _, EvmTranscript<_, _, _, _>, _>(
+                &self.agg_params,
+                &vk,
+                AccumulatorStrategy::new(&self.params),
+                &verify_circuit_instance2,
+                &mut transcript,
+            )
+            .unwrap(),
+        ))
     }
 
     pub fn verify_target_circuit_proof<C: TargetCircuit>(
         &mut self,
         proof: &TargetCircuitProof,
     ) -> anyhow::Result<()> {
-        let instances: Vec<Vec<Vec<u8>>> = serde_json::from_reader(&proof.instance[..])?;
-        let instances = deserialize_fr_matrix(instances);
-
+        let instances = proof.snark.instances.clone();
         let instance_slice = instances.iter().map(|x| &x[..]).collect::<Vec<_>>();
 
         let verifier_params = self.params.verifier_params();
 
-        let mut transcript = PoseidonRead::<_, _, Challenge255<_>>::init(&proof.proof[..]);
+        let mut transcript: PoseidonTranscript<_,_> = TranscriptReadBuffer::<_, G1Affine, _>::init(proof.snark.proof.as_slice());
         let strategy = SingleStrategy::new(verifier_params);
 
         let vk = self.target_circuit_vks.entry(C::name()).or_insert_with(|| {
-            let circuit = C::empty();
+            let circuit = C::dummy_inner_circuit();
             keygen_vk(&self.params, &circuit)
                 .unwrap_or_else(|_| panic!("failed to generate {} vk", C::name()))
         });
 
-        verify_proof::<_, VerifierGWC<_>, _, _, _>(
+        verify_proof::<_, VerifierSHPLONK<_>, _, _, _>(
             verifier_params,
             vk,
             strategy,
@@ -131,5 +126,12 @@ impl Verifier {
             &mut transcript,
         )?;
         Ok(())
+    }
+
+
+    /// Verifies the proof with EVM byte code.
+    /// Panics if verification fails.
+    pub fn evm_verify(bytecode: Vec<u8>, instances: Vec<Vec<Fr>>, proof: Vec<u8>) {
+        evm_verify(bytecode, instances, proof)
     }
 }
