+++ zkevm/tests/mock_plonk.rs
@@ -1,0 +1,164 @@
+//! A module for Mock Plonk circuit.
+//!
+use halo2_proofs::halo2curves::bn256::Fr;
+use halo2_proofs::{
+    circuit::{Layouter, SimpleFloorPlanner, Value},
+    plonk::{Advice, Circuit, Column, ConstraintSystem, Error, Fixed, Instance},
+    poly::Rotation,
+};
+use rand::{RngCore};
+use snark_verifier_sdk::CircuitExt;
+use zkevm::circuit::TargetCircuit;
+use zkevm_circuits::witness;
+
+pub struct MockPlonkCircuit {
+    pub circuit: StandardPlonk
+}
+
+impl TargetCircuit for MockPlonkCircuit {
+    /// The actual inner circuit that implements Circuit trait.
+    type Inner = StandardPlonk;
+
+    /// Name tag of the circuit.
+    /// This tag will be used as a key to index the circuit.
+    /// It is therefore important that the name is unique.
+    fn name() -> String {
+        "standard plonk".into()
+    }
+
+    /// Generate a dummy circuit with an empty trace.
+    /// This is useful for generating vk and pk.
+    fn dummy_inner_circuit() -> Self::Inner
+    where
+        Self: Sized,
+    {   
+        StandardPlonk(Fr::zero())
+    }
+
+
+    /// Build the inner circuit and the instances from the witness block
+    fn from_witness_block(
+        _witness_block: &witness::Block<Fr>,
+    ) -> anyhow::Result<(Self::Inner, Vec<Vec<Fr>>)>
+    where
+        Self: Sized,
+    {
+        unimplemented!()
+    }
+}
+
+#[derive(Clone, Copy)]
+pub struct StandardPlonkConfig {
+    a: Column<Advice>,
+    b: Column<Advice>,
+    c: Column<Advice>,
+    q_a: Column<Fixed>,
+    q_b: Column<Fixed>,
+    q_c: Column<Fixed>,
+    q_ab: Column<Fixed>,
+    constant: Column<Fixed>,
+    #[allow(dead_code)]
+    instance: Column<Instance>,
+}
+
+impl StandardPlonkConfig {
+    fn configure(meta: &mut ConstraintSystem<Fr>) -> Self {
+        let [a, b, c] = [(); 3].map(|_| meta.advice_column());
+        let [q_a, q_b, q_c, q_ab, constant] = [(); 5].map(|_| meta.fixed_column());
+        let instance = meta.instance_column();
+
+        [a, b, c].map(|column| meta.enable_equality(column));
+
+        meta.create_gate(
+            "q_a·a + q_b·b + q_c·c + q_ab·a·b + constant + instance = 0",
+            |meta| {
+                let [a, b, c] = [a, b, c].map(|column| meta.query_advice(column, Rotation::cur()));
+                let [q_a, q_b, q_c, q_ab, constant] = [q_a, q_b, q_c, q_ab, constant]
+                    .map(|column| meta.query_fixed(column, Rotation::cur()));
+                let instance = meta.query_instance(instance, Rotation::cur());
+                Some(
+                    q_a * a.clone()
+                        + q_b * b.clone()
+                        + q_c * c
+                        + q_ab * a * b
+                        + constant
+                        + instance,
+                )
+            },
+        );
+
+        StandardPlonkConfig {
+            a,
+            b,
+            c,
+            q_a,
+            q_b,
+            q_c,
+            q_ab,
+            constant,
+            instance,
+        }
+    }
+}
+
+#[derive(Clone, Copy, Default, Debug)]
+pub struct StandardPlonk(Fr);
+
+impl StandardPlonk {
+    pub fn rand<R: RngCore>(mut rng: R) -> Self {
+        Self(Fr::from(rng.next_u32() as u64))
+    }
+}
+
+impl CircuitExt<Fr> for StandardPlonk {
+    fn num_instance(&self) -> Vec<usize> {
+        vec![1]
+    }
+
+    fn instances(&self) -> Vec<Vec<Fr>> {
+        vec![vec![self.0]]
+    }
+}
+
+impl Circuit<Fr> for StandardPlonk {
+    type Config = StandardPlonkConfig;
+    type FloorPlanner = SimpleFloorPlanner;
+
+    fn without_witnesses(&self) -> Self {
+        Self::default()
+    }
+
+    fn configure(meta: &mut ConstraintSystem<Fr>) -> Self::Config {
+        meta.set_minimum_degree(4);
+        StandardPlonkConfig::configure(meta)
+    }
+
+    fn synthesize(
+        &self,
+        config: Self::Config,
+        mut layouter: impl Layouter<Fr>,
+    ) -> Result<(), Error> {
+        layouter.assign_region(
+            || "",
+            |mut region| {
+                region.assign_advice(|| "", config.a, 0, || Value::known(self.0))?;
+                region.assign_fixed(|| "", config.q_a, 0, || Value::known(-Fr::one()))?;
+                region.assign_advice(|| "", config.a, 1, || Value::known(-Fr::from(5u64)))?;
+                for (idx, column) in (1..).zip([
+                    config.q_a,
+                    config.q_b,
+                    config.q_c,
+                    config.q_ab,
+                    config.constant,
+                ]) {
+                    region.assign_fixed(|| "", column, 1, || Value::known(Fr::from(idx as u64)))?;
+                }
+                let a = region.assign_advice(|| "", config.a, 2, || Value::known(Fr::one()))?;
+                a.copy_advice(|| "", &mut region, config.b, 3)?;
+                a.copy_advice(|| "", &mut region, config.c, 4)?;
+
+                Ok(())
+            },
+        )
+    }
+}
