+++ zkevm/tests/snark_verifier_api.rs
@@ -1,7 +1,5 @@
-// This is essentially a same test as snark-verifier/evm-verifier
-use application::StandardPlonk;
-use halo2_proofs::halo2curves::bn256::Fr;
 use halo2_proofs::poly::commitment::Params;
+use mock_plonk::MockPlonkCircuit;
 use rand::SeedableRng;
 use rand_xorshift::XorShiftRng;
 use snark_verifier::loader::halo2::halo2_ecc::halo2_base::utils::fs::gen_srs;
@@ -9,146 +7,19 @@ use snark_verifier_sdk::evm::{evm_verify, gen_evm_proof_shplonk, gen_evm_verifie
 use snark_verifier_sdk::halo2::aggregation::AggregationCircuit;
 use snark_verifier_sdk::CircuitExt;
 use snark_verifier_sdk::{gen_pk, halo2::gen_snark_shplonk};
+use zkevm::io::serialize_vk;
 use std::path::Path;
-
-mod test_util;
 use test_util::init;
+use zkevm::prover::{Prover, TargetCircuitProof};
+use zkevm::verifier::Verifier;
+use mock_plonk::StandardPlonk;
 
-mod application {
-    use super::*;
-    use halo2_proofs::{
-        circuit::{Layouter, SimpleFloorPlanner, Value},
-        plonk::{Advice, Circuit, Column, ConstraintSystem, Error, Fixed, Instance},
-        poly::Rotation,
-    };
-    use rand::RngCore;
-    use snark_verifier_sdk::CircuitExt;
+mod mock_plonk;
+mod test_util;
 
-    #[derive(Clone, Copy)]
-    pub struct StandardPlonkConfig {
-        a: Column<Advice>,
-        b: Column<Advice>,
-        c: Column<Advice>,
-        q_a: Column<Fixed>,
-        q_b: Column<Fixed>,
-        q_c: Column<Fixed>,
-        q_ab: Column<Fixed>,
-        constant: Column<Fixed>,
-        #[allow(dead_code)]
-        instance: Column<Instance>,
-    }
-
-    impl StandardPlonkConfig {
-        fn configure(meta: &mut ConstraintSystem<Fr>) -> Self {
-            let [a, b, c] = [(); 3].map(|_| meta.advice_column());
-            let [q_a, q_b, q_c, q_ab, constant] = [(); 5].map(|_| meta.fixed_column());
-            let instance = meta.instance_column();
-
-            [a, b, c].map(|column| meta.enable_equality(column));
-
-            meta.create_gate(
-                "q_a·a + q_b·b + q_c·c + q_ab·a·b + constant + instance = 0",
-                |meta| {
-                    let [a, b, c] =
-                        [a, b, c].map(|column| meta.query_advice(column, Rotation::cur()));
-                    let [q_a, q_b, q_c, q_ab, constant] = [q_a, q_b, q_c, q_ab, constant]
-                        .map(|column| meta.query_fixed(column, Rotation::cur()));
-                    let instance = meta.query_instance(instance, Rotation::cur());
-                    Some(
-                        q_a * a.clone()
-                            + q_b * b.clone()
-                            + q_c * c
-                            + q_ab * a * b
-                            + constant
-                            + instance,
-                    )
-                },
-            );
-
-            StandardPlonkConfig {
-                a,
-                b,
-                c,
-                q_a,
-                q_b,
-                q_c,
-                q_ab,
-                constant,
-                instance,
-            }
-        }
-    }
-
-    #[derive(Clone, Default)]
-    pub struct StandardPlonk(Fr);
-
-    impl StandardPlonk {
-        pub fn rand<R: RngCore>(mut rng: R) -> Self {
-            Self(Fr::from(rng.next_u32() as u64))
-        }
-    }
-
-    impl CircuitExt<Fr> for StandardPlonk {
-        fn num_instance(&self) -> Vec<usize> {
-            vec![1]
-        }
-
-        fn instances(&self) -> Vec<Vec<Fr>> {
-            vec![vec![self.0]]
-        }
-    }
-
-    impl Circuit<Fr> for StandardPlonk {
-        type Config = StandardPlonkConfig;
-        type FloorPlanner = SimpleFloorPlanner;
-
-        fn without_witnesses(&self) -> Self {
-            Self::default()
-        }
-
-        fn configure(meta: &mut ConstraintSystem<Fr>) -> Self::Config {
-            meta.set_minimum_degree(4);
-            StandardPlonkConfig::configure(meta)
-        }
-
-        fn synthesize(
-            &self,
-            config: Self::Config,
-            mut layouter: impl Layouter<Fr>,
-        ) -> Result<(), Error> {
-            layouter.assign_region(
-                || "",
-                |mut region| {
-                    region.assign_advice(|| "", config.a, 0, || Value::known(self.0))?;
-                    region.assign_fixed(|| "", config.q_a, 0, || Value::known(-Fr::one()))?;
-                    region.assign_advice(|| "", config.a, 1, || Value::known(-Fr::from(5u64)))?;
-                    for (idx, column) in (1..).zip([
-                        config.q_a,
-                        config.q_b,
-                        config.q_c,
-                        config.q_ab,
-                        config.constant,
-                    ]) {
-                        region.assign_fixed(
-                            || "",
-                            column,
-                            1,
-                            || Value::known(Fr::from(idx as u64)),
-                        )?;
-                    }
-                    let a = region.assign_advice(|| "", config.a, 2, || Value::known(Fr::one()))?;
-                    a.copy_advice(|| "", &mut region, config.b, 3)?;
-                    a.copy_advice(|| "", &mut region, config.c, 4)?;
-
-                    Ok(())
-                },
-            )
-        }
-    }
-}
-
+// This is essentially a same test as snark-verifier/evm-verifier
 #[test]
-fn test_snark_verifier_api() {
+fn test_snark_verifier_sdk_api() {
     std::env::set_var("VERIFY_CONFIG", "./configs/example_evm_accumulator.config");
     let k = 8;
     let k_agg = 21;
@@ -204,4 +75,117 @@ fn test_snark_verifier_api() {
 
     println!("finished bytecode generation");
     evm_verify(deployment_code, instances, proof)
+}
+
+
+// A partial integration test.
+// The inner snark proofs are generated from a mock circuit 
+// instead of the trace files.
+#[test]
+fn test_aggregation_api() {
+    std::env::set_var("VERIFY_CONFIG", "./configs/example_evm_accumulator.config");
+
+    init();
+    let num_snarks = 3;
+    let k = 8;
+    let k_agg = 21;
+    let seed = [0u8; 16];
+    let mut rng = XorShiftRng::from_seed(seed);
+
+    let circuit = StandardPlonk::rand(&mut rng);
+    let params_outer = gen_srs(k_agg);
+    let params_inner = {
+        let mut params = params_outer.clone();
+        params.downsize(k);
+        params
+    };
+
+    let prover = Prover::from_params_and_seed(params_inner, params_outer, seed);
+
+    // ====================================================
+    // A whole aggregation procedure takes the following steps
+    // 1. instantiation the parameters and the prover
+    // 2. convert block traces into inner circuit proofs, a.k.a. SNARKs
+    // 3. build an aggregation circuit proof
+    // 4. generate bytecode for evm to verify aggregation circuit proof
+    // 5. validate the proof with evm bytecode
+    // ====================================================
+    //
+    // 1. instantiation the parameters and the prover
+    //
+    let k = 8;
+    let k_agg = 21;
+    let seed = [0u8; 16];
+    let mut rng = XorShiftRng::from_seed(seed);
+
+    // notice that k < k_agg which is not necessary the case in practice
+    let params_outer = gen_srs(k_agg);
+    let params_inner = {
+        let mut params = params_outer.clone();
+        params.downsize(k);
+        params
+    };
+    log::info!("loaded parameters for degrees {} and {}", k, k_agg);
+
+    //
+    // 2. convert block traces into inner circuit proofs, a.k.a. SNARKs
+    //
+    // Note:
+    // we do not have traces for testing so here we simply assume that we have already
+    // obtained 3 inner circuits proofs for some dummy circuit
+    let pk_inner = gen_pk(&params_inner, &circuit, None);
+    let vk_inner = pk_inner.get_vk();
+    let vk_inner_bytes = serialize_vk(vk_inner);
+
+    let snarks = (0..num_snarks)
+        .map(|_| {
+            gen_snark_shplonk(
+                &params_inner,
+                &pk_inner,
+                circuit.clone(),
+                &mut rng,
+                None::<String>,
+            )
+        })
+        .collect::<Vec<_>>();
+    log::info!("finished inner circuit snark generation");
+
+    // sanity check: the inner proof is correct
+    let mut verifier = Verifier::new(params_inner, params_outer.clone(), None);
+    for i in 0..num_snarks {
+        let proof = TargetCircuitProof {
+            name: "MockPlonk".into(),
+            snark: snarks[i].clone(),
+            vk: vk_inner_bytes.clone(),
+            num_of_proved_blocks: 0,
+            total_num_of_blocks: 0,
+        };
+        verifier.verify_target_circuit_proof::<MockPlonkCircuit>(&proof).unwrap();
+    }
+    log::info!("sanity check: inner circuit snark are correct");
+
+    // 3. build an aggregation circuit proof
+    let agg_circuit = AggregationCircuit::new(&params_outer, snarks, &mut rng);
+    let pk_outer = gen_pk(&params_outer, &agg_circuit, None);
+
+    let instances = agg_circuit.instances();
+    let proof = gen_evm_proof_shplonk(
+        &params_outer,
+        &pk_outer,
+        agg_circuit.clone(),
+        instances.clone(),
+        &mut rng,
+    );
+    log::info!("finished aggregation generation");
+
+    // 4. generate bytecode for evm to verify aggregation circuit proof
+    let deployment_code = prover.create_evm_verifier_bytecode(
+        &agg_circuit,
+        pk_outer.get_vk(),
+    );
+    log::info!("finished byte code generation");
+
+    // 5. validate the proof with evm bytecode
+    Verifier::evm_verify(deployment_code, instances, proof);
+    log::info!("end to end test completed");
 }
